+++ resources/views/bootstrap-5/components/table/cell.blade.php
@@ -1,3 +1,5 @@
-<td {{ $attributes }}>
+@props(['customAttributes' => []])
+
+<td {{ $attributes->merge($customAttributes) }}>
     {{ $slot }}
 </td>