+++ resources/views/tailwind/components/table/row-columns.blade.php
@@ -1,6 +1,10 @@
 @foreach($columns as $column)
     @if ($column->isVisible())
-        <x-livewire-tables::table.cell>
+        <x-livewire-tables::table.cell
+            :class="method_exists($this, 'setTableDataClass') ? $this->setTableDataClass($column, $row) : ''"
+            :id="method_exists($this, 'setTableDataId') ? $this->setTableDataId($column, $row) : ''"
+            :customAttributes="method_exists($this, 'setTableDataAttributes') ? $this->setTableDataAttributes($column, $row) : []"
+        >
             @if ($column->asHtml)
                 {{ new \Illuminate\Support\HtmlString($column->formatted($row)) }}
             @else